fix: contentful allows broken links, lets avoid creating broken refs
