+++ src/test/helpers/index.ts
@@ -4,3 +4,4 @@ export * from './contentfulEditorControlFactory'
 export * from './contentfulEditorInterfaceFactory'
 export * from './contentfulEnvironmentLinkFactory'
 export * from './contentfulSpaceLinkFactory'
+export * from './parse'
